// src/types/ui.type.ts
export enum SpinnerSize {
  SMALL = "sm",
  MEDIUM = "md",
  LARGE = "lg",
  EXTRA_LARGE = "xl",
}

export enum LoadingAction {
  SAVING = "saving",
  SEARCHING = "searching",
  PROCESSING = "processing",
  LOADING = "loading",
  UPLOADING = "uploading",
  DOWNLOADING = "downloading",
  SYNCING = "syncing",
}